Added a ICashHolder interface and more
I added the ICashHolder interface, which will be implemented by all
classes that will store and type of money or credit. I also added a
whole lot of new utilities in the CashUtil class. Of course, I also
implemented the ICashHolder interface within the classes Card and
Wallet. I created a new class called Transaction, but I'm not sure
whether it will be of use as of now.